feat(workspace-cap): migrate profile module (slice 10 of N preparatory)
ADR-152 section 3 F3.4 preparatory slice. Verbatim port of profile.rs
(1145 lines including 16 tests, no internal use crate dependencies) from
desktop-client/ironclaw/src/profile.rs into dasclaw_workspace_cap::profile.
ironclaw side becomes a 6-line re-export shim, matching the slice 9 pattern.
All 6 direct call sites inside Workspace are switched to the cap path,
removing Workspace's direct dependency on crate::profile. The 2 remaining
call sites in setup/profile_evolution.rs resolve through the shim.

Closes #710
Refs #674
